refactoring domains

+++ src/main/java/it/istat/is2/app/bean/InputFormBean.java
@@ -40,7 +40,7 @@ public class InputFormBean implements Serializable {
 	private String tipo;
 	private String labelFile;
 	private String labelCodeRule;
-	private Integer fileType;
+	private Long fileType;
 	private String covariate;
 	private String predizioni;
 	private int numeroCampi;
